pyproject.toml
[tool.poetry]
name = "insta-bot"
version = "0.1.0"
description = ""
authors = ["Ivan Kovtun <ivan.kovtun@capgemini.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
bs4 = "^0.0.2"
instagrapi = "^2.0.3"
diffusers = "^0.26.2"
transformers = "^4.37.2"
accelerate = "^0.27.0"
pillow = "^10.2.0"
opencv-python = "^4.9.0.80"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
